Fix indentation and improve TPCx-HS download logic

+++ benchmarks/Spark/conf/configure.sh
@@ -47,16 +47,18 @@ fi
 if [ "$GEN_TPCX_HS" == "true" ]; then
 	export TPCX_HS_JAR=$SPARK_BENCH_DIR/$SPARK_TPCX_HS_JAR_NAME
 
-	# Download TPCx-HS jar file
-	URL=https://bdev.des.udc.es/dist/tpcx-hs
-	m_echo "Downloading $SPARK_TPCX_HS_JAR_NAME from $URL"
+	if [ ! -f $TPCX_HS_JAR ]; then
+		# Download TPCx-HS jar file
+		URL=https://bdev.des.udc.es/dist/tpcx-hs
+		m_echo "Downloading $SPARK_TPCX_HS_JAR_NAME from $URL"
 
-    wget -q -O $TPCX_HS_JAR $URL/$SPARK_TPCX_HS_JAR_NAME
+    	wget -q -O $TPCX_HS_JAR $URL/$SPARK_TPCX_HS_JAR_NAME
 
-    if [ $? != 0 ]; then
-		rm $TPCX_HS_JAR >& /dev/null
-		m_exit "Error when downloading $SPARK_TPCX_HS_JAR_NAME"
-    fi
-else
-	m_echo "Using $TPCX_HS_JAR"
+    	if [ $? != 0 ]; then
+			rm $TPCX_HS_JAR >& /dev/null
+			m_exit "Error when downloading $SPARK_TPCX_HS_JAR_NAME"
+   		fi
+	else
+		m_echo "Using $TPCX_HS_JAR"
+	fi
 fi
